LayoutTests/fast/regex/overflow-expected.txt

This test checks expressions with alternative lengths of appox. 2^31.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regexp1.exec('') is null
PASS regexp2.exec('') is null
PASS successfullyParsed is true

TEST COMPLETE
